% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawn_boot.R
\name{pawn_ci}
\alias{pawn_ci}
\title{Computation of confidence intervals}
\usage{
pawn_ci(data)
}
\arguments{
\item{data}{The output of the \code{pawn_indices} function.}
}
\value{
A data table.
}
\description{
Computation of confidence intervals
}
\examples{
\donttest{# Create matrix:
A <- randtoolbox::sobol(n = 1000, dim = 8)
# Compute model output:
Y <- sobol_Fun(A)
Ti <- pawn_generic(data = A, Y = Y, n = 10, test = median, R = 50 )
# Compute confidence intervals:
pawn_ci(Ti)}
}
