% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawn_boot.R
\name{pawn_dummy}
\alias{pawn_dummy}
\title{Computation of the PAWN index of a dummy parameter}
\usage{
pawn_dummy(Y, n, R)
}
\arguments{
\item{Y}{Numeric vector including the unconditional model output.}

\item{n}{Integer, number of conditioning intervals.}

\item{R}{Integer, number of bootstrap replicas.}
}
\value{
A matrix with the bootstrap replicas.
}
\description{
The function computes and bootstraps the PAWN index of a dummy parameter
following the approach by \insertCite{Pianosi2018;textual}{pawnr}.
}
\examples{
# Create matrix:
A <- randtoolbox::sobol(n = 1000, dim = 8)
# Compute model output:
Y <- sobol_Fun(A)
# Compute PAWN index of a dummy parameter:
dummy <- pawn_dummy(Y, n = 10, R = 10)
}
\references{
\insertAllCited{}
}
