% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawn_generic.R
\name{pawn_generic}
\alias{pawn_generic}
\title{Computation of PAWN indices (generic approach)}
\usage{
pawn_generic(data, Y, n, test, R)
}
\arguments{
\item{data}{A data frame or matrix with the sample matrix.}

\item{Y}{Numeric vector, model output.}

\item{n}{Integer, number of conditioning intervals.}

\item{test}{Summary statistic (i.e. \code{mean}, \code{median}, \code{max}).}

\item{R}{Integer, number of bootstrap replicas.}
}
\value{
A data table with the bootstrap PAWN indices of each model input.
}
\description{
This function computes and bootstraps PAWN indices following the
"generic approach" of \insertCite{Pianosi2018;textual}{pawnr}.
}
\details{
The bootstrap is conducted without replacement, as recommended by
\insertCite{KhorashadiZadeh2017;textual}{pawnr}.
}
\examples{
\donttest{# Create matrix:
A <- randtoolbox::sobol(n = 100, dim = 8)
# Compute model output:
Y <- sobol_Fun(A)
T <- pawn_generic(data = A, Y = Y, n = 10, test = median, R = 50 )}
}
\references{
\insertAllCited{}
}
