% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawn_matrix.R
\name{pawn_matrix}
\alias{pawn_matrix}
\title{Creation of the matrices to compute PAWN (tailored approach)}
\usage{
pawn_matrix(N, Nc, n, k)
}
\arguments{
\item{N}{Numeric, sample size of the unconditional model output}

\item{Nc}{Numeric, sample size of the conditional model output}

\item{n}{Numeric, number of conditioning intervals}

\item{k}{Numeric, number of parameters}
}
\value{
A matrix
}
\description{
Creates the matrices to compute the model output and then calculate
the PAWN sensitivity indices of the model inputs following the
'tailored approach'. The matrix for the unconditional model output
is created via Sobol' quasi-random number sequences.
}
\examples{
pawn_matrix(N = 10,
Nc = 5,
n = 3,
k = 3)
}
