% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawn_plot.R
\name{plot_pawn}
\alias{plot_pawn}
\title{Plot PAWN indices}
\usage{
plot_pawn(data)
}
\arguments{
\item{data}{The output of the \code{pawn_ci} function.}
}
\value{
A ggplot object
}
\description{
Plot PAWN indices
}
\examples{
\donttest{# Create matrix:
A <- randtoolbox::sobol(n = 1000, dim = 8)
# Compute model output:
Y <- sobol_Fun(A)
T <- pawn_generic(data = A, Y = Y, n = 10, test = median, R = 50 )
T.ci <- pawn_ci(T)
plot_pawn(T.ci)}
}
